requirements.txt
Django==1.6.7
Pillow==2.7.0
South==1.0.2
argparse==1.2.1
cov-core==1.15.0
coverage==3.7.1
django-nose==1.3
django-stdimage==1.2.1
docutils==0.12
nose==1.3.4
progressbar==2.2
py==1.4.26
pytest==2.6.4
pytest-cov==1.8.1
selenium==2.44.0
wsgiref==0.1.2
django-ajax-upload-widget==0.5.5
mock==1.0.1
